backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
celery==5.5.2
redis==6.0.0
python-dotenv==1.0.1
django-filter==24.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
setuptools==69.2.0
django-celery-email==3.0.0
coreapi==2.3.3 